Baby Sweettooth theme options
This code adds a theme options page to a theme called Baby Sweettooth that I wrote in 2010. I used this theme options code on numerous themes, some of which I sold through the theme shop I ran, Museum Themes. The code was taken from a blog post on WPShout, which took it from something else, and it was passed around for a long time before standards were set about best practices for theme options.